backend/requirements.txt
keybert==0.7.0
matplotlib==3.6.2
Flask==2.2.2
Flask-Cors==3.0.10
gensim==4.2.0
newsapi==0.1.1
newsapi_python==0.2.6
nltk==3.7
numpy==1.23.4
pandas==1.5.1
rake_nltk==1.0.6
scikit_fuzzy==0.4.2
scikit_learn==1.1.3
spacy==3.4.2
unicode_slugify==0.1.5
